strutil: move IDAsHex/IDFromHex to new pkg
